requirements.txt
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cx-Oracle==7.2.3
decorator==4.4.1
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.8
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.0
jupyterlab-server==1.0.6
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.2
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.17.4
openpyxl==3.0.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
plotly==4.5.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Pygments==2.5.2
pylint==2.5.3
pymongo==3.10.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.6.0
requests==2.22.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.13.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
yfinance==0.1.54
